refactor(re-arranging files to avoid unecessary react dependency in most of the library): moved all
the only method that relies on react are the ones provided by the createApiWithCss and the only
method that depends on createApiWithCss is the flushChunks. By separating flushChunks from the rest
of the methods I'm making this library a little bit more independent and more  like a universal
solution to the chunks problems

26
